--- examples/clearml/toy_ex.py
@@ -1,14 +1,0 @@
-import pandas as pd
-from clearml import Task, StorageManager
-
-#task = Task.init(
-#            project_name="a.p.chikov/test/clearml_basic",
-#           task_name="experiment_artefacts",
-#        )
-
-#metrics = pd.DataFrame({
-#    "a": [1,2,3],
-#    "b": [3,4,5]
-#})
-#task.logger.report_table(title="Metrics", series="Lol", iteration=0, table_plot=metrics)
-#task.logger.report_table(title="Metrics", series="Lol", iteration=1, table_plot=metrics)
